fix: marketplace page refactored
