+++ src/components/MarketplaceGrid.vue
@@ -1,57 +1,19 @@
 <template>
-  <v-container v-if="!loading && !error">
+  <v-container>
     <v-row>
       <v-col v-for="product in productList" :key="product.id">
         <MarketplaceCard :product="product" />
       </v-col>
     </v-row>
   </v-container>
-
-  <ProgressCircular v-if="loading" />
-
-  <AlertCard
-    v-if="error"
-    :on-close="handleCloseClick"
-    :text="error.message"
-    title="Connection error!"
-    type="error"
-  />
 </template>
 
 <script lang="ts" setup>
-import AlertCard from '@/components/AlertCard.vue'
 import MarketplaceCard from '@/components/MarketplaceCard.vue'
-import ProgressCircular from '@/components/ProgressCircular.vue'
-import { useNavigationDrawerStore } from '@/stores/navigationDrawer'
 import { useProductListStore } from '@/stores/productList'
-import { Route } from '@/types/route'
 import { storeToRefs } from 'pinia'
-import { onBeforeMount, ref } from 'vue'
 
 const productListStore = useProductListStore()
 
 const { productList } = storeToRefs(productListStore)
-const { getProducts } = productListStore
-
-const { selectRoute } = useNavigationDrawerStore()
-
-const loading = ref(false)
-
-const error = ref<Error | null>(null)
-
-const changeLoading = () => {
-  loading.value = !loading.value
-}
-
-const handleError = (msg: Readonly<string>) => {
-  error.value = new Error(msg)
-}
-
-const handleCloseClick = () => {
-  selectRoute(Route.HOMEPAGE)
-}
-
-onBeforeMount(async () => {
-  await getProducts(handleError, changeLoading)
-})
 </script>
